pyproject.toml
[project]
name = "git-agent"
version = "0.1.0"
description = "An A2A-compliant agent for git operations using Claude and MCP"
authors = [
    { name = "Bobby Battista" },
]
dependencies = [
    "python-a2a[anthropic,mcp]",
    "pydantic-ai",
    "click",
    "python-dotenv"
]
requires-python = ">=3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["git_agent"] 